Issue Fixes, Simplified Logic
I fixed the issues of the last pull request and gave output when the program is launched incorrectly without sufficient arguments. I made it (slightly) more memory efficient. The Controller class now also handles the button combination. The --controller parameter, being a helpless guessing game, is still for the more advanced user. I changed the name 'runningEmulator' to 'runProgram' to make it more universal to programs in general.

In its current state, it's pretty much done. It could use some error checking to see if paths are valid and possibly a few other things but in terms of core functionality, it's done. Each controller separately checks the button combination as opposed to combining all the button presses of every controller. When the user is missing necessary parameters, QuitBit outputs a (colorful!) guide explaining the use of the argument and an example of each.

The multithreading of the orignal commit may have to be re-implemented if the hanging issue reoccurs, but this is pretty much done, unless someone gets a good idea.